refactor: moved the server.js compiler to own module in order to allow compilers to share logic
affects: oc-server-compiler, oc-template-handlebars-compiler, oc-template-jade-compiler
