+++ packages/oc-server-compiler/README.md
@@ -1,0 +1,3 @@
+# oc-server-compiler
+
+OC module to compile server.js
