+++ packages/oc-server-compiler/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "oc-server-compiler",
+  "version": "1.0.0",
+  "description": "OC-Server-Compiler",
+  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/opencomponents/base-templates.git"
+  },
+  "bugs": {
+    "url": "https://github.com/opencomponents/base-templates/issues"
+  },
+  "homepage": "https://github.com/opencomponents/base-templates#readme",
+  "keywords": [
+    "oc",
+    "opencomponents"
+  ],
+  "author": {
+    "name": "Matteo Figus",
+    "email": "matteofigus@gmail.com"
+  },
+  "license": "MIT",
+  "dependencies": {
+    "async": "^2.4.1",
+    "fs-extra": "^4.0.0",
+    "oc-hash-builder": "1.0.1",
+    "oc-webpack": "^1.0.6"
+  },
+  "files": [
+    "index.js"
+  ]
+}
