+++ packages/oc-template-handlebars-compiler/package.json
@@ -31,6 +31,7 @@
     "oc-get-unix-utc-timestamp": "1.0.1",
     "oc-hash-builder": "1.0.1",
     "oc-minify-file": "^1.0.3",
+    "oc-server-compiler": "^1.0.0",
     "oc-template-handlebars": "^6.0.2",
     "oc-templates-messages": "^1.0.1",
     "oc-view-wrapper": "1.0.2",
