+++ packages/oc-template-jade-compiler/package.json
@@ -34,6 +34,7 @@
     "oc-get-unix-utc-timestamp": "1.0.1",
     "oc-hash-builder": "1.0.1",
     "oc-minify-file": "^1.0.3",
+    "oc-server-compiler": "^1.0.0",
     "oc-template-jade": "6.0.1",
     "oc-templates-messages": "^1.0.1",
     "oc-view-wrapper": "1.0.2",
